Corrections of somes bugs!

+++ src/utils/lst_utils.c
@@ -6,7 +6,7 @@
 /*   By: rgobet <rgobet@student.42angouleme.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/03 15:01:08 by rgobet            #+#    #+#             */
-/*   Updated: 2024/07/06 16:14:52 by rgobet           ###   ########.fr       */
+/*   Updated: 2024/07/09 08:22:54 by rgobet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,8 @@ static void	free_node_resources(t_redirection *node, t_vars *vars)
 		free(node->name_outfile);
 	if (node->name_infile)
 		free(node->name_infile);
-	if (node->e_position == HERE && vars->cmd_index != 0)
+	if (node->e_position == HERE && vars->cmd_index != 0
+		&& vars->child != 0)
 		unlink(node->file_heredoc);
 	if (node->limiter)
 		free(node->limiter);
